backend/requirements.txt
fastapi
uvicorn
openai
google-generativeai
groq
python-dotenv
